Use the hvm repo types and fns; Switch to u64 nodes

+++ src/ast/mod.rs
@@ -11,7 +11,7 @@ pub struct Name(pub String);
 
 // TODO: Accept other number types
 #[derive(Debug, PartialEq, Eq, Clone, Copy, Shrinkwrap)]
-pub struct Number(pub u32);
+pub struct Number(pub u64);
 
 #[derive(Debug, PartialEq, Eq, Clone, Copy, Shrinkwrap, Hash, PartialOrd, Ord)]
-pub struct DefId(pub u32);
+pub struct DefId(pub u64);
